move crypto config into crypto package
